requirements.txt
# RAG System Dependencies
# Core ML/NLP
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.2
accelerate>=0.25.0

# LangChain ecosystem
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Vector stores and retrieval
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU available
chromadb>=0.4.0
rank-bm25>=0.2.2

# Embeddings
openai>=1.6.0
tiktoken>=0.5.0

# Evaluation
ragas>=0.1.0
datasets>=2.16.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0

# Development
jupyter>=1.0.0
ipykernel>=6.28.0

